refactor(canvas): colocate chat page body hooks
