+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/hooks/__tests__/useChatPageJumpNavigation.test.tsx
@@ -2,7 +2,7 @@
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it, vi } from 'vitest';
-import { useChatPageJumpNavigation } from './useChatPageJumpNavigation';
+import { useChatPageJumpNavigation } from '../useChatPageJumpNavigation';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
