+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/hooks/__tests__/useChatPageNewSession.test.tsx
@@ -2,7 +2,7 @@
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it, vi } from 'vitest';
-import { useChatPageNewSession } from './useChatPageNewSession';
+import { useChatPageNewSession } from '../useChatPageNewSession';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
