+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/hooks/__tests__/useChatPageTargetContext.test.tsx
@@ -2,9 +2,9 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import type { ChatContextSnapshot } from '../../../agent-chat/target';
-import { useChatPageTargetContext } from './useChatPageTargetContext';
+import { I18nProvider } from '../../../../../i18n';
+import type { ChatContextSnapshot } from '../../../../../agent-chat/target';
+import { useChatPageTargetContext } from '../useChatPageTargetContext';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
