+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/hooks/useChatPageNewSession.ts
@@ -1,8 +1,8 @@
 import { useCallback, useEffect, useRef } from 'react';
-import { scopeSessionStoreId } from '../../../../../shared/agent-chat';
-import type { AgentNewSessionResult, AgentScope } from '../types';
-import { clearChatComposerDraft } from './chatComposerDraftStore';
-import { restoreComposerFocusAfterRender } from '../utils/focusRecovery';
+import { scopeSessionStoreId } from '../../../../../../shared/agent-chat';
+import type { AgentNewSessionResult, AgentScope } from '../../types';
+import { clearChatComposerDraft } from '../../hooks/chatComposerDraftStore';
+import { restoreComposerFocusAfterRender } from '../../utils/focusRecovery';
 
 interface Options {
   agentScope: AgentScope;
