+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/hooks/useChatPagePendingSession.ts
