+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/hooks/useStableSessionRail.ts
@@ -1,11 +1,11 @@
 import { useMemo, useRef } from 'react';
-import { scopeSessionStoreId } from '../../../../../shared/agent-chat';
-import { conversationKeyId } from '../../../../../shared/conversation-runtime';
-import { useI18n } from '../../../i18n';
-import type { AgentSessionInfo } from '../../../types';
-import type { UnifiedSession } from '../ChatSessionsRail';
-import type { AgentScope, OtherWorkspaceSession, WorkspaceOption } from '../types';
-import type { ConversationCompletionStatus } from '../../../agent-chat/runtime/conversationCompletionStore';
+import { scopeSessionStoreId } from '../../../../../../shared/agent-chat';
+import { conversationKeyId } from '../../../../../../shared/conversation-runtime';
+import { useI18n } from '../../../../i18n';
+import type { AgentSessionInfo } from '../../../../types';
+import type { UnifiedSession } from '../../ChatSessionsRail';
+import type { AgentScope, OtherWorkspaceSession, WorkspaceOption } from '../../types';
+import type { ConversationCompletionStatus } from '../../../../agent-chat/runtime/conversationCompletionStore';
 
 interface UseStableSessionRailOptions {
   agentScope: AgentScope;
