+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatPageBody/index.tsx
@@ -1,35 +1,35 @@
 import { useCallback, useEffect, useMemo, type KeyboardEventHandler, type ReactNode } from 'react';
-import type { CanvasNode } from '../../types';
-import { useRightDock, useRightDockState } from '../dock/RightDock/context';
-import { isDockContentTabVisible, toggleFullPageDockContentTabs } from '../dock/RightDock/dock-content-tabs';
-import type { SettingsSection } from '../settings/Settings';
-import './ChatPage.css';
-import './ChatPanel.css';
-import type { UnifiedSession } from './ChatSessionsRail';
-import { ChatView } from './ChatView';
-import { SessionBackBar, type SessionBackEntry } from './SessionBackBar';
-import { useChatComposerStateKeyed } from './hooks/useChatComposerStateKeyed';
-import { isExternalOnlyRoleMessage } from '../../agent-chat/mentions/roleMentionItems';
-import { useAppShell } from '../shell/AppShellProvider';
-import type { AgentNewSessionResult, AgentScope, WorkspaceOption } from './types';
-import { useI18n } from '../../i18n';
-import { isImeComposing } from '../../utils/ime';
-import type { ChatContextSnapshot, ChatExecutionPolicy, ChatTarget } from '../../agent-chat/target';
-import { chatScopeId } from './chatScope';
-import { useRegisterChatTarget } from '../../agent-chat/target/useRegisterChatTarget';
-import { ChatConversationStatus } from './ChatConversationStatus';
+import type { CanvasNode } from '../../../types';
+import { useRightDock, useRightDockState } from '../../dock/RightDock/context';
+import { isDockContentTabVisible, toggleFullPageDockContentTabs } from '../../dock/RightDock/dock-content-tabs';
+import type { SettingsSection } from '../../settings/Settings';
+import '../ChatPage.css';
+import '../ChatPanel.css';
+import type { UnifiedSession } from '../ChatSessionsRail';
+import { ChatView } from '../ChatView';
+import { SessionBackBar, type SessionBackEntry } from '../SessionBackBar';
+import { useChatComposerStateKeyed } from '../hooks/useChatComposerStateKeyed';
+import { isExternalOnlyRoleMessage } from '../../../agent-chat/mentions/roleMentionItems';
+import { useAppShell } from '../../shell/AppShellProvider';
+import type { AgentNewSessionResult, AgentScope, WorkspaceOption } from '../types';
+import { useI18n } from '../../../i18n';
+import { isImeComposing } from '../../../utils/ime';
+import type { ChatContextSnapshot, ChatExecutionPolicy, ChatTarget } from '../../../agent-chat/target';
+import { chatScopeId } from '../chatScope';
+import { useRegisterChatTarget } from '../../../agent-chat/target/useRegisterChatTarget';
+import { ChatConversationStatus } from '../ChatConversationStatus';
 import { useChatPageTargetContext } from './hooks/useChatPageTargetContext';
 import { useChatPageJumpNavigation } from './hooks/useChatPageJumpNavigation';
 import { useChatPageSessionRail } from './hooks/useChatPageSessionRail';
-import { useScopeRunningSessions } from '../../agent-chat/sessions/useScopeRunningSessions';
+import { useScopeRunningSessions } from '../../../agent-chat/sessions/useScopeRunningSessions';
 import { useChatPagePendingSession } from './hooks/useChatPagePendingSession';
-import { useSubmitDomReviewComments } from './hooks/useSubmitDomReviewComments';
-import { submitQuickAction } from './hooks/submitQuickAction';
-import { ChatPageRail, ChatPageTopbar } from './ChatPageNavigationChrome';
-import { scopeSessionStoreId } from '../../../../shared/agent-chat';
-import { buildChatPageDockTabRefs } from './utils/chatPageDockTabs';
+import { useSubmitDomReviewComments } from '../hooks/useSubmitDomReviewComments';
+import { submitQuickAction } from '../hooks/submitQuickAction';
+import { ChatPageRail, ChatPageTopbar } from '../ChatPageNavigationChrome';
+import { scopeSessionStoreId } from '../../../../../shared/agent-chat';
+import { buildChatPageDockTabRefs } from '../utils/chatPageDockTabs';
 import { useChatPageNewSession } from './hooks/useChatPageNewSession';
-import { useConversationCompletionNotices } from '../../agent-chat/sessions/useConversationCompletionNotices';
+import { useConversationCompletionNotices } from '../../../agent-chat/sessions/useConversationCompletionNotices';
 export interface ChatPageBodyProps {
   agentScope: AgentScope;
   /** Context inherited from the visible target that opened this page. */
