+++ apps/canvas-workspace/src/renderer/src/components/chat/__tests__/ChatNewSession.focus.test.tsx
@@ -4,7 +4,7 @@ import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { I18nProvider } from '../../../i18n';
 import { ChatSessionsRail } from '../ChatSessionsRail';
-import { useChatPageSessionRail } from '../hooks/useChatPageSessionRail';
+import { useChatPageSessionRail } from '../ChatPageBody/hooks/useChatPageSessionRail';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
